added properties for connection, homework moved to spec. package
