+++ dad.properties
@@ -1,0 +1,5 @@
+adressDB=localhost
+portDB=5432
+nameDB=dad
+userName=postgres
+userPass=root
