+++ src/icmit/dad/laborators/lab1/homework/DBWork.java
@@ -3,12 +3,10 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package icmit.dad.laborators.lab1;
+package icmit.dad.laborators.lab1.homework;
 
-import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.sql.Connection;
 import java.sql.DriverManager;
